// src/common/types.ts
import { TFunction } from "react-i18next";
export interface ContainerProps {
  border?: boolean;
  children: React.ReactNode;
}

export interface ButtonProps {
  color?: string;
  name?: string;
  children: React.ReactNode;
  onClick?: () => void;
}

export interface SvgIconProps {
  src: string;
  width: string;
  height: string;
}

export interface InputProps {
  name: string;
  placeholder: string;
  t: TFunction;
  type?: string;
  value?: string;
  onChange: (
    event:
      | React.ChangeEvent<HTMLInputElement>
      | React.ChangeEvent<HTMLTextAreaElement>
  ) => void;
}
export interface validateProps {
  yourName: string;
  yourPhone: string;
  yourEmail: string;
  yourAddress: string;

  report: string;
  date: string;
  CSC: string;
  location: string;

  suspectName: string;
  suspectPostion: string;
  suspectPhone: string;

  File: string[];
}